+++ src/main/java/springschool/ranking/record/service/dto/partial/PartialRateRecordDto.java
@@ -7,7 +7,7 @@ public class PartialRateRecordDto extends PartialRecordDto {
 
     private final double rate;
 
-    public PartialRateRecordDto(Long studentId, long score, int year, int semester, double rate) {
+    public PartialRateRecordDto(Long studentId, double score, int year, int semester, double rate) {
         super(studentId, score, year, semester);
         this.rate = rate;
     }
